Remove superfluous config system
The current config system is superfluous at best. Everyone who wants
to change Kint settings does so by changing the public static Kint
properties.

In fact, I'd go so far as to say anyone using the config system as
it exists before this patch is producing unmaintainable code, as
they are polluting the Kint directory as opposed to setting the
settings in their own code.

There should be no need to explain why the config file is bad
code when you consider a composer environment...