Renderer system overhaul
This new system leaves more of the implementation up to the
renderer subclasses, which in turn gives them more freedom.

As a demonstration, all of the header special cases inside
Kint_Renderer_Rich (Depth limit, recursion, no argument etc)
have been moved to plugins (Behaviour which wasn't possible
from plugins before)

This isn't obligatory for renderer implementations but it's
nice and clean and paves the way for new features.