rich.js: Major rewrite
Mostly cleanup, abstraction, sensibility renaming.

Also ensures click focus and keyboard focus are synced, so if you
click on an element and start keyboard mode it will start focused
on that one. Likewise, if you are in keyboard mode and switch
focus the keyboard caret system will be updated correctly.

The kint object has been named kintRich and is available for
outside use, providing an API of sorts for plugins using the
pre_render_sources plugin callbacks