Kint_Parser_Plugin: Add getCleanArray() to help plugins with validation
Parser plugins may want to operate on the raw data instead of the
Kint_Object when altering things. This is a helper function to
return a copy of an array without a marker in it.

Fixes broken backtrace plugin since 76354df6fc714f60a

DO NOT pass an array that has had it's marker removed back
into the parser, it will result in an extra recursion.