Kint_Parser_Plugin_Blacklist
Creating the blacklist plugin leads to an interesting problem:
How do we blacklist something?

If we simply mark the Kint_Object with a blacklist hint, there's
no guarantee the renderer won't ignore the warning or leave it
unimplemented and disclose sensitive information anyway!

So we do need to destroy all the data in the Kint_Object inside
the parser. But this leads to a new problem: What do we do about
plugins that come after the blacklist and add information to the
Kint_Object.

There are two ways of handling this:

1. Ensure the blacklist is always at the bottom of the list.

    While this works, it provides new problems every time a
    third party plugin shows up, as it has to dilligently make
    sure it's not upsetting a fragile order of operations.

    This fragility can also exasperate existing fragile inter-
    plugin dependancies to an unacceptable amount.
2. Give the blacklist the ability to force an end to plugin
   rendering on return.

    This adds painful complexity to an otherwise gloriously
    simple procedure, but it's a neccesity to pull off the
    desired behaviour without constantly hacking around it.