Kint_Parser_Plugin_Base64: Try not to parse normal words
I noticed it has a tendancy to try to parse words like "page" or
"password" since they are technically valid base64.

This should prevent that since it's unlikely any normal string
that's valid base64 (Multiple of 4 length, no spaces etc) will
be over 16 chars.

Additionally, we don't place the base64 representation at the
front of the list unless it's over 50 chars - just to be safe.

A string of over 50 chars that matches base64 will probably be
something we should decode, but under 50 may well just be strange
input. (Take the valid color "lightgoldenrodyellow" for instance)