SimpleXML: Bugfixes for odd data structures
Apparently you can access comments as nodes in SimpleXML, so if you
have a node with the actual name "comment" it's going to mess
everything up royally.

Here we manually check which elements there are and guarantee the
access paths will be correct by forcing an index on each element
(so the path will be `->node[0]` at least instead of `->node`)

This includes a few visual aids so nodes with text contents will
show their short value since it removes some of the hierarchy.