Performance optimization
Grinding some caches for faster performance!

This is a huge commit with tons of little changes - some making as
little as 0.1% difference in the run time (Practically placebo)

Still, the total effect is 45% - 55% shorter runtime so it was worth it!

It's still slower than kint 1 - though this is because it presents
more information. It's more than 10% faster per byte than kint 1.

If performance is still a problem after this, try blacklisting
heavy objects with the blacklist plugin or writing your own
plugin for more special case behaviour.