Rewrite parser plugin system
This rewrite ensures that the parser itself is a much more
separate entity to the whole Kint system, and allows plugins
to more accurately choose when they're run (Including before
the base parse)

Instead of a large static array containing all of the plugins,
you call addPlugin on the parser with a plugin object when you
want to add it to the pool.

addPlugin then calls getTypes and getTriggers on the plugin to
decide which pools the plugin should be added to, and setParser
to ensure the plugin has access to the parser while running.

Since the plugin methods getTypes and getTriggers combined return
the circumstances the plugin should be run, there's less chance
for misconfiguration and we can remove some runtime checks.

This does mean we have to add the plugins manually, in this case
moving our plugin instantiation into the Kint class.

We can squeeze some more speed out of this in future by
maintaining separate plugin lists for renderers that don't
use all of the plugins outputs anyway.

There's the downside of a performance hit from constructing all
the parser plugins ahead of time, but we can mitigate that with
the ad-hoc rule that plugins used from the Kint helper class
should be cached and still work when attached to a new parser,
since anything going through the Kint class pretty much guarantees
we won't be running multiple parsers at the same time.

Separate systems using the parser and renderer classes individually
should use the addPlugin() API at will.