Add single file compilation
Well that was easier than I thought it would be!

The only real change to core kint code is splitting the init,
and ensuring the Rich and Plain renderers can take strings in
addition to callables (So we can bake in the CSS and JS files)

While the previous require list was carefully picked to be in
the right order to prevent dependency issues, the new one is
built statically and in *theory* should work in any order
(Since they're all defined in the same file). In practice it's
freaking out when some classes aren't in the right order, and
not when others aren't. Might have something to do with abstract.

Anyway, CI says hitting the file list with a sort() is good
enough, and if we ever add a file that breaks this behavior
CI will complain.