Single file compilation manually
The classpreloader turned out to be more trouble than it was
really worth. Manually stripping the namespaces to support 5.1
already made it essentially concatenation, but worse still it
had a tendancy to change escaped sequences into literal
unprintable characters. This is simpler, with smaller output
files and it works just as well.