Use namespaces
Well this is a big one. PSR-4!

Since we can now use namespaces this also has some implications
on naming conventions. Certain classes like Kint_Object can't
simply be renamed to Kint\Object since Object is an illegal
class name as of PHP 7.2.

Similarly, bundling parent classes in a namespace of the same
name makes more sense now since you can skip a use statement
on them. For example, by naming it Kint\Parser\Parser instead
of Kint\Parser you can skip a use for all the plugins in the
Kint\Parser namespace that need to use it.

No namespace separators in code. All classes should be declared
up front in the use statements making it easy to find a classes
dependencies and dependants.

To prevent constant "use as" aliases for classes with conflicting
names in different namespaces, the actual class names have been
renamed to be more specific such as "MicrotimePlugin",
"MicrotimeObject", "MicrotimeRepresentation". Since the
namespaces will be removed in the use statements the total length
of the class in code is likely to still be smaller now.

Renamed SourceParser to CallFinder since in the end it's a
very specialized class and the new name suits it better.

Since most of the classes are going to be renamed anyway, this
is a perfect opportunity to do some BC-breaking cleanups. The
rich renderer now has interfaces for its plugins. While it
doesn't check them at runtime for performance reasons it's a
quick way to ensure function signatures are correct.

Thanks to the interfaces we now have seperate methods for
rendering an object and a tab in the rich renderer, so we
can merge the separate color plugins into a single plugin
that handles rendering both the object and the tab.

Similarly the DOMNode and DOMIterator parser plugins have
been merged into a single DOMDocumentPlugin class.