Value: Require name in constructor
With name being required this lets us skip checking it in various
places like the renderer, and psalm is much happier with the
situation. The downside is that you can't have "Unnamed" values.

Previously, the only place unnamed values would occur was when the
CallFinder failed to find an input, or the input was simplistic
and the facade elided the name completely. In these cases I've
opted to replace the missing name with "literal", "expression",
or a numeric identifier that is invalid in PHP like the access
path already uses (Starting at $0 and going $1, $2, etc.)