Parser: Use ReflectionReference to check for references
Vastly simplifies the reference checking process. Interestingly,
because _everything_ in PHP is refcounted something that isn't a
reference will become one if some other code takes a reference to
it.

This is a problem when you're looping over an array by reference
because everything will be a reference _only at the time you
check_ and go back to a normal variable after the loop so that
was a fun heisenbug to find.

So I loop over copies instead and pass it to the recursive call by
key instead.