CallFinder: T_NEW is no longer an operator
"Operator" tokens in callfinder are used to determine whether a
parameter is an "expression" that needs to be wrapped in parens
in an access path.

As of PHP 8.4 you no longer need parens around an initialization.

Unfortunately, only for T_NEW and T_CLASS (As called in anonymous
classes `new class {}`) are the only non-operators that should
keep whitespace around them, so we have to add a separate check
for them to preserve spaces.